+++ FiapCloudGamesAPI/Models/BibliotecaDoJogador.cs
@@ -4,12 +4,12 @@ using FiapCloudGamesAPI.Entidades;
 
 namespace FiapCloudGamesAPI.Models
 {
-    public class BibliotecaDoJogador(int idUsuario, int idJogo, string criadoPor) : EntidadeBase(criadoPor)
+    public class BibliotecaDoJogador(long idUsuario, long idJogo, string criadoPor) : EntidadeBase(criadoPor)
     {
-        public int IdUsuario { get; set; } = idUsuario;
+        public long IdUsuario { get; set; } = idUsuario;
         public Usuario Usuario { get; set; }
 
-        public int IdJogo { get; set; } = idJogo;
+        public long IdJogo { get; set; } = idJogo;
         public Jogo Jogo { get; set; }
